Add files via upload
In this exercise, you will produce an implementation of the Ladner-Fischer† parallel prefix sum algorithm using the C++ thread libraryLinks to an external site.. The interior nodes of the tree in the algorithm will be implemented using an array (like a heap). Write a Heaper base class as shown in the lecture slides and extend that for your own SumHeap class. The data array will be in a std::vector<int> typedef'd to be called Data.

†The Ladner-Fischer algorithm is the one presented in the textbook in Chapter 1 and discussed in lecture in week 1 and week 2.

You are allowed to make the number of elements in the array constrained to be a power of 2. There is extra credit for relaxing this requirement such that the algorithm still works correctly.

The recursive adder pass must be parallelized for the top four levels of the tree by forking a thread, using std::async, at each of those levels but done within the same thread for levels below that (total of 16 threads). Similarly, the prefix summation pass must also be done in parallel in the same fashion.

In your testing, use smaller arrays, of course, but for your final submission use an array of 67,108,864 elements (226) which on CS1 should run in about 1000ms without optimization (with -O2, my 8-thread solution runs in 260ms). If you are handing in a submission that includes the extra credit, then make your data array 100,000,000 elements and note that you are turning in the extra credit and what your strategy is.

Your program should all be in a single file, hw2.cpp.

It should include a test outside the timing loop. To satisfy this requirement, you may use the following code verbatim:

const int N = 1<<26;  // FIXME must be power of 2 for now

int main() {
    Data data(N, 1);  // put a 1 in each element of the data array
    data[0] = 10;
    Data prefix(N, 1);

    // start timer
    auto start = chrono::steady_clock::now();

    SumHeap heap(&data);
    heap.prefixSums(&prefix);

    // stop timer
    auto end = chrono::steady_clock::now();
    auto elpased = chrono::duration<double,milli>(end-start).count();

    int check = 10;
    for (int elem: prefix)
        if (elem != check++) {
            cout << "FAILED RESULT at " << check-1;
            break;
        }
    cout << "in " << elpased << "ms" << endl;
    return 0;
}
Guidance
Passing methods to async
For the thread forks in your recursive routines, you will likely want to call std::async with a method instead of a function as demonstrated in the example programs. To do this, since it is not a static method, you need to pass in this as the first argument. My call to async in my recursive method calcSum, looks like this, for instance:

auto handle = async(launch::async, &SumHeap::calcSum, this, left(i), level+1);
Thread Allocations
Take a look at Fig. 5.19(a) in the textbook to see how we are handling thread allocation of each of the subtrees in our pair-wise sum tree. You may want to keep track of the level of the recursion by adding in a parameter to your calcSum and calcPrefix methods so they know when to stop forking threads. Alternatively, you could write a level(i) method for your Heaper class.

Two-Pass Algorithm
The prefix sum algorithm is a two-pass algorithm.

The pair-wise sum pass. This is done during the construction of the SumHeap object and must be done in parallel (using 16 threads as noted above).
SumHeap heap(&data);
The prefix sum pass. This is done within the prefixSums method and must also be done in parallel. The base case of this recursion writes the prefix sum values into the output array.
heap.prefixSums(&prefix);