Add files via upload
Pipelined Bitonic Sorter
Implement a bitonic sorter pipelining program in Java.

You will have a sorting pipeline which continuously takes in a stream of arrays of floating point numbers (i.e., the double datatype in Java) and outputs the arrays in sorted order. For our testing, the pipeline will be hooked up to a random array generator on the front end and on the back end it will do a check to look for any out of order elements and then discard the array. The BitonicPipeline class has the main method that hooks everything up and sets them in motion. We'll run it for a fixed amount of time, say 10 seconds, to demonstrate it working and gather timing measurements.

Sequential Benchmark
But before we embark on the BitonicPipeline program, the first iteration of the project is to have a sequential bitonic sorting net as a benchmark as well as the testing apparatus. Once that is complete, we can measure its throughput and use that as a benchmark for our parallelization. This program is in a separate class, BitonicSequential (provided), that is structured similarly to BitonicPipeline, but does not use any threads and does not use any SynchronousQueue objects. It should report the number of arrays processed per second for a given output array size, N (defined in your program somewhere). Like BitonicPipeline will, it uses System.currentTimeMillis to measure wall-clock time and then stop after ten seconds have elapsed.

The instructor provides the BitonicSequential class, as well as the RandomArrayGenerator class, but the student is responsible for the BitonicStage class which does a bitonic sort of its bitonic sequence input. Both the BitonicPipeline and BitonicSequential use this bitonic sorter as a building block.

Parallelizing the Pipelined Bitonic Sort
Then in the next iteration which , the student will parallelize the parts of the network into a pipeline and measure the speedup achieved. For this, there will be the driver class, BitonicPipeline, as well as the basic bitonic sorter (taking a bitonic sequence as input), BitonicStage, and finally an initial stage, StageOne, that takes an unordered array and sorts it.

The pipeline will consist of seven pipeline threads (in addition to the main thread and four random array generator threads). Four StageOne threads (threads I through IV) will do the initial processing of the array, each owning one quarter of the elements. Two BitonicStage threads (threads V and VI) will do the processing at the next layer, each handling one half the array. The final additional BitonicStage thread (thread VII) will do the final bitonic sort of the whole array. So threads I, III, V, and VII sort their section in ascending order, and threads II, IV, VI sort their sections in descending order. Threads I - IV (StageOne) are allowed to use any sort algorithm, such as the library sort from Arrays class. This will be slightly faster than using a bitonic sorter internally there. The higher BitonicStage threads (V - VII) must use bitonic sort to take advantage of the bitonic series coming in.

Each thread will wait for each of its input section(s), then proceed with its task, and post the output for the next thread in the pipeline. We will use the SynchronousQueueLinks to an external site. class in Java to handle these pipelined data buses (the SynchronousQueue objects work just like the full/empty variables in the textbook). The random generator threads continuously produce random double arrays that are consumed by threads I - IV. The main thread after starting up all the threads, will keep consuming the output of Thread VII and check it. It also tracks and reports statistics on throughput.

The size of the whole array is determined by a static constant N in the main class. N must be a power of 2.

Here's a diagram of what the worker threads would be responsible for in the case where N = 16 (of course, we expect N to be much larger than that once everything is working):

Thread Allocation

Here's the suggested class breakdown:

BitonicPipeline class has main and hosts the main thread. (You could break this down into a BitonicPipeline class and a separate test harness class.)
RandomArrayGenerator (provided) class implements Runnable and is parameterized by the size of the array (one quarter the size of the final array that is output by Thread VII) and its SynchronousQueue output object.
StageOne (provided) class implements Runnable and is the class for threads I - IV. Parameterized by size (from the size of the array in the input queue), input and output queues. This stage always outputs ascending sorted arrays.
BitonicStage class implements Runnable and is the class for threads V - VII. Parameterized by size, input queues (2) and output queue. You can have it assume that the two arrays from the input queues are both in ascending sort order and then have it invert the second one to make the composite of the first and the inverted second a fully-fledged bitonic sequence.